+++ asyncgit/src/sync/branch/rename.rs
@@ -37,7 +37,7 @@ mod test {
 
 		create_branch(repo_path, "branch1").unwrap();
 
-		checkout_branch(repo_path, "refs/heads/branch1").unwrap();
+		checkout_branch(repo_path, "branch1").unwrap();
 
 		assert_eq!(
 			repo.branches(None)